# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Digital Workplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Projects ────────────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Attendance ────────────────────────────────────────────────────────────
...
  /analytics/activity:
    get:
      operationId: getActivityFeed
      tags: [analytics]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /analytics/productivity:
    get:
      operationId: getProductivityAnalytics
      tags: [analytics]
      summary: Get productivity analytics
      parameters:
        - name: period
          in: query
          schema: { type: string, enum: [week, month, quarter] }
      responses:
        "200":
          description: Productivity data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductivityAnalytics"

  # ── Storage ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]

    # ── Projects ────────────────────────────────────────────────────────────
...
    Announcement:
      type: object
      required: [id, title, content, priority, authorId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        pinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        pinned:
          type: boolean

    # ── Attendance ────────────────────────────────────────────────────────────
...
    ActivityItem:
      type: object
      required: [id, type, description, actorName, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [task_created, task_completed, project_created, document_uploaded, leave_requested, announcement_posted, user_joined]
        description:
          type: string
        actorName:
          type: string
        actorAvatar:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProductivityAnalytics:
      type: object
      required: [period, taskCompletionRate, avgTasksPerDay, topPerformers, weeklyTrend]
      properties:
        period:
          type: string
        taskCompletionRate:
          type: number
        avgTasksPerDay:
          type: number
        topPerformers:
          type: array
          items:
            $ref: "#/components/schemas/UserProductivityStat"
        weeklyTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
    UserProductivityStat:
      type: object
      required: [userId, userName, completedTasks, avgHoursPerTask]
      properties:
        userId:
          type: integer
        userName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        completedTasks:
          type: integer
        avgHoursPerTask:
          type: number
    TrendPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
...